Twenty commits: a typed calculator evaluator, quicklink arguments moved
from their own screen into a header accessory, extension background
refresh and search-bar dropdowns, Escape walking back through screens,
and Raycast script-command and quicklink imports.

Ten conflicts, no divergence re-litigated. The retired
QuicklinkArgumentSession takes its two previews and their fixture with
it, and PreviewChrome now injects CustomCommandArgumentSession.
Replace-on-import sits on upstream's extracted addImportedQuicklinks.
RootPaletteView was taken whole from upstream's new stateObservers /
keyHandlers split, with the scope chip, the suggestion and Linear
observers, launcherChord and the scale tokens re-applied onto it. The
new argument strip and the quicklink preview glyph now scale, and
palette-navigation-test compiles the scope sources it needs.

+++ Tinycast/Features/QuickActions/Model/TextDiffEngine.swift
@@ -8,7 +8,7 @@ enum TextDiffEngine: Sendable {
         case deleted(String)
     }
 
-    /// The LCS matrix is quadratic, so an unbounded diff of a long selection asks for gigabytes.
+    /// The matrix is quadratic, so an unbounded diff asks for gigabytes; a cell must fit `UInt16`.
     static let maxTokens = 4_000
 
     static func diff(original: String, modified: String) -> [Chunk] {
@@ -61,9 +61,9 @@ enum TextDiffEngine: Sendable {
         return tokens
     }
 
-    private static func longestCommonSubsequence(_ old: [String], _ new: [String]) -> [[Int]] {
+    private static func longestCommonSubsequence(_ old: [String], _ new: [String]) -> [[UInt16]] {
         var table = Array(
-            repeating: Array(repeating: 0, count: new.count + 1), count: old.count + 1)
+            repeating: Array(repeating: UInt16(0), count: new.count + 1), count: old.count + 1)
         for i in 0..<old.count {
             for j in 0..<new.count {
                 table[i + 1][j + 1] =
